Refactor: split core.py into focused modules
Break the monolithic core.py (1,225 lines) into 6 focused modules:
- filters.py: row-level filtering primitives (DTE, OTM%, delta, signals)
- pricing.py: fill price calculation, slippage models, P&L computation
- evaluation.py: entry/exit matching pipeline, _calls/_puts filters
- calendar.py: calendar/diagonal spread evaluation and exit matching
- output.py: result formatting and statistical grouping
- core.py: orchestration layer with _strategy_engine, _process_strategy,
  _process_calendar_strategy, and backward-compatible re-exports

No public API changes. All existing imports from core.py (strategies.py,
datafeeds.py, tests/test_rules.py) continue to work via re-exports.
